requirements.txt
Django==3.9.x
djangorestframework==3.12.4